Pill-Detection/requirements.txt
anyio==3.6.2
aws-shell==0.2.2
awscli==1.27.133
boto3==1.26.133
botocore==1.29.133
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.4
configobj==5.0.8
docutils==0.16
fastapi==0.95.0
filelock==3.11.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imutils==0.5.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
pandas==2.0.0
Pillow==9.5.0
prompt-toolkit==1.0.18
pyasn1==0.5.0
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==5.4.1
requests==2.28.2
rsa==4.7.2
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
